+++ stream-chat-android-ui-components/src/main/kotlin/com/getstream/sdk/chat/viewmodel/messages/MessageListViewModel.kt
@@ -52,6 +52,7 @@ import io.getstream.chat.android.offline.extensions.cancelEphemeralMessage
 import io.getstream.chat.android.offline.extensions.getRepliesAsState
 import io.getstream.chat.android.offline.extensions.loadMessageById
 import io.getstream.chat.android.offline.extensions.loadNewerMessages
+import io.getstream.chat.android.offline.extensions.loadNewestMessages
 import io.getstream.chat.android.offline.extensions.loadOlderMessages
 import io.getstream.chat.android.offline.extensions.setMessageForReply
 import io.getstream.chat.android.offline.extensions.watchChannelAsState
@@ -337,6 +338,7 @@ public class MessageListViewModel(
             deletedMessageVisibility = deletedMessageVisibility,
             messageFooterVisibility = messageFooterVisibility,
             messagePositionHandlerProvider = ::messagePositionHandler,
+            endOfNewMessages = channelState.endOfNewerMessages.asLiveData(),
         )
         _reads.addSource(channelState.reads.asLiveData()) { _reads.value = it }
         _loadMoreLiveData.addSource(channelState.loadingOlderMessages.asLiveData()) { _loadMoreLiveData.value = it }
@@ -385,6 +387,7 @@ public class MessageListViewModel(
             deletedMessageVisibility = deletedMessageVisibility,
             messageFooterVisibility = messageFooterVisibility,
             messagePositionHandlerProvider = ::messagePositionHandler,
+            endOfNewMessages = MutableLiveData(true)
         )
         threadListData?.let { tld ->
             messageListData?.let { mld ->
@@ -693,6 +696,26 @@ public class MessageListViewModel(
     }
 
     /**
+     * When the user clicks the scroll to bottom button we need to take the user to the bottom of the newest
+     * messages. If the messages are not loaded we need to load them first and then scroll to the bottom of the
+     * list.
+     */
+    public fun scrollToBottom(scrollToBottom: () -> Unit) {
+        if (_mode.value is Mode.Thread || messageListData?.value?.areNewestMessagesLoaded == true) {
+            scrollToBottom()
+        } else {
+            chatClient.loadNewestMessages(cid, DEFAULT_MESSAGES_LIMIT).enqueue { result ->
+                if (result.isSuccess) {
+                    scrollToBottom()
+                } else {
+                    val error = result.error()
+                    logger.e { "Could not load newest messages. Cause: ${error.cause?.message}" }
+                }
+            }
+        }
+    }
+
+    /**
      * Sets the date separator handler which determines when to add date separators.
      * By default, a date separator will be added if the difference between two messages' dates is greater than 4h.
      *
@@ -769,9 +792,9 @@ public class MessageListViewModel(
         currentMode.run {
             when (this) {
                 is Mode.Normal -> {
-                    messageListData?.loadingMoreChanged(true)
+                    messageListData?.loadingMoreOldMessagesChanged(true)
                     chatClient.loadOlderMessages(cid, DEFAULT_MESSAGES_LIMIT).enqueue {
-                        messageListData?.loadingMoreChanged(false)
+                        messageListData?.loadingMoreOldMessagesChanged(false)
                     }
                 }
                 is Mode.Thread -> threadLoadMore(this)
@@ -784,10 +807,10 @@ public class MessageListViewModel(
      */
     private fun onBottomEndRegionReached(baseMessageId: String?) {
         if (baseMessageId != null) {
-            messageListData?.loadingMoreChanged(true)
+            messageListData?.loadingMoreNewMessagesChanged(true)
             chatClient.loadNewerMessages(cid, baseMessageId, DEFAULT_MESSAGES_LIMIT)
                 .enqueue { result ->
-                    messageListData?.loadingMoreChanged(false)
+                    messageListData?.loadingMoreNewMessagesChanged(false)
                 }
         } else {
             logger.e { "There's no base message to request more message at bottom of limit" }
@@ -800,17 +823,17 @@ public class MessageListViewModel(
      * @param threadMode Current thread mode.
      */
     private fun threadLoadMore(threadMode: Mode.Thread) {
-        threadListData?.loadingMoreChanged(true)
+        threadListData?.loadingMoreOldMessagesChanged(true)
         if (threadMode.threadState != null) {
             chatClient.getRepliesMore(
                 messageId = threadMode.parentMessage.id,
                 firstId = threadMode.threadState.oldestInThread.value?.id ?: threadMode.parentMessage.id,
                 limit = DEFAULT_MESSAGES_LIMIT,
             ).enqueue {
-                threadListData?.loadingMoreChanged(false)
+                threadListData?.loadingMoreOldMessagesChanged(false)
             }
         } else {
-            threadListData?.loadingMoreChanged(false)
+            threadListData?.loadingMoreOldMessagesChanged(false)
             logger.w { "Thread state must be not null for offline plugin thread load more!" }
         }
     }