+++ stream-chat-android-ui-components/src/main/kotlin/io/getstream/chat/android/ui/message/list/MessageListView.kt
@@ -824,6 +824,13 @@ public class MessageListView : ConstraintLayout {
     }
 
     /**
+     * Scrolls the message list to the bottom.sl
+     */
+    public fun scrollToBottom() {
+        scrollHelper.scrollToBottom()
+    }
+
+    /**
      * Set a custom layout manager for MessageListView. This can be used to change orientation of messages.
      *
      * @param layoutManager
@@ -1179,7 +1186,8 @@ public class MessageListView : ConstraintLayout {
                     scrollHelper.onMessageListChanged(
                         isThreadStart = isThreadStart,
                         hasNewMessages = listItem.hasNewMessages,
-                        isInitialList = isOldListEmpty && filteredList.isNotEmpty()
+                        isInitialList = isOldListEmpty && filteredList.isNotEmpty(),
+                        areNewestMessagesLoaded = listItem.areNewestMessagesLoaded
                     )
 
                     buffer.active()
@@ -1622,6 +1630,15 @@ public class MessageListView : ConstraintLayout {
     }
 
     /**
+     * Sets the handler used when the user interacts with [ScrollButtonView].
+     *
+     * @param handler The handler to use.
+     */
+    public fun setOnScrollToBottomHandler(handler: OnScrollToBottomHandler) {
+        this.scrollHelper.setScrollToBottomHandler(handler)
+    }
+
+    /**
      * Handles the selected [messageAction].
      *
      * @param messageAction The newly selected action.
@@ -1854,6 +1871,10 @@ public class MessageListView : ConstraintLayout {
 
     public fun interface ModeratedMessageOptionHandler {
         public fun onModeratedMessageOptionSelected(message: Message, moderatedMessageOption: ModeratedMessageOption)
+    }
+
+    public fun interface OnScrollToBottomHandler {
+        public fun onScrollToBottom()
     }
     //endregion