+++ stream-chat-android-ui-components/src/main/kotlin/io/getstream/chat/android/ui/message/list/viewmodel/MessageListViewModelBinding.kt
@@ -87,6 +87,7 @@ public fun MessageListViewModel.bindView(
         }
     }
     view.setReplyMessageClickListener { messageId -> onEvent(MessageListViewModel.Event.ShowMessage(messageId)) }
+    view.setOnScrollToBottomHandler { scrollToBottom { view.scrollToBottom() } }
 
     ownCapabilities.observe(lifecycleOwner) {
         view.setOwnCapabilities(it)